remove fp16

+++ finetune_pl.py
@@ -128,5 +128,5 @@ if __name__ == "__main__":
     model = LM()
 
     trainer = Trainer(
-        num_tpu_cores=8, progress_bar_refresh_rate=1, precision=16)
+        num_tpu_cores=8, progress_bar_refresh_rate=1)
     trainer.fit(model)
